refactor(agents): share agent price bottom-navigation bar
Extract AgentPriceBottomNavigation (price + divider + trailing-action slot)
from the now near-identical AgentBottomNavigation and JobBottomNavigation,
which differed only by their trailing action.
